+++ packages/qiskit-algo/test/functional.js
@@ -11,7 +11,7 @@
 
 const assert = require('assert');
 
-const devs = require('..');
+const algo = require('..');
 const { version } = require('../package');
 
 // TODO: Use utils.difference instead.
@@ -19,10 +19,10 @@ function multiIncludes(text, values) {
   return values.every(val => text.includes(val));
 }
 
-describe('devs:api', () => {
+describe('algo:api', () => {
   it('should include all documented items', () =>
     assert.ok(
-      multiIncludes(Object.keys(devs), [
+      multiIncludes(Object.keys(algo), [
         'version',
         'random',
         'factor',
@@ -31,5 +31,5 @@ describe('devs:api', () => {
     ));
 });
 
-describe('devs:version', () =>
-  it('should be correct', () => assert.equal(devs.version, version)));
+describe('algo:version', () =>
+  it('should be correct', () => assert.equal(algo.version, version)));