+++ CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `@qiskit/qiskit-sim`: Add print method to Circuit API documentation
 - `@qiskit/qiskit-sim`: Add 'add' method to Circuit class
 
+> - 🐛 **Fixed**: for any bug fixes.
+
+- `@qiskit/qiskit-sim`: Fix connection bug in printCircuit
+
 ## [0.8.0] - 2019-04-25
 
 ### 🎉 Added