+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.9.0] - 2019-05-13
+
 ### ✏️ Changed
 
 - `@qiskit/qiskit-sim`: Use spread operator in addGate